Move metadata/helper to tools/metadatahelper
