+++ pkg/tools/metadatahelper/int_meta_data.gen.go
@@ -18,7 +18,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package helper
+package metadatahelper
 
 import (
 	"context"
