+++ pkg/tools/metadatahelper/meta_data.template.go
@@ -14,7 +14,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package helper
+package metadatahelper
 
 import (
 	"context"
